refactor: split recoil files
